# PERF: periodarr_to_dt64arr