--- a/asv_bench/benchmarks/tslibs/period.py
+++ b/asv_bench/benchmarks/tslibs/period.py
@@ -2,9 +2,14 @@
 Period benchmarks that rely only on tslibs.  See benchmarks.period for
 Period benchmarks that rely on other parts fo pandas.
 """
-from pandas import Period
+
+import numpy as np
+
+from pandas._libs.tslibs.period import Period, periodarr_to_dt64arr
 
 from pandas.tseries.frequencies import to_offset
+
+from .tslib import _sizes
 
 
 class PeriodProperties:
@@ -68,3 +73,34 @@ class PeriodConstructor:
 
     def time_period_constructor(self, freq, is_offset):
         Period("2012-06-01", freq=freq)
+
+
+class TimePeriodArrToDT64Arr:
+    params = [
+        _sizes,
+        [
+            1000,
+            1011,  # Annual - November End
+            2000,
+            2011,  # Quarterly - November End
+            3000,
+            4000,
+            4006,  # Weekly - Saturday End
+            5000,
+            6000,
+            7000,
+            8000,
+            9000,
+            10000,
+            11000,
+            12000,
+        ],
+    ]
+    param_names = ["size", "freq"]
+
+    def setup(self, size, freq):
+        arr = np.arange(size, dtype="i8")
+        self.i8values = arr
+
+    def time_periodarray_to_dt64arr(self, size, freq):
+        periodarr_to_dt64arr(self.i8values, freq)