@jbrockmendel 2020-07-08
> can you comment on the slowdowns on the particular cases, e.g. 4000, 5000, I guess we don't care about these as much

As mentioned in the OP, i dont understand why we see slowdowns this big.  It should just be one extra `if freq >= FR_DAY` check, shouldn't make a major difference in the non-shortcircuited cases