@jbrockmendel 2020-07-08
i think i might be able to fastpath some more cases, will give that a shot